services/qrcode/requirements.txt
aiogram==2.25.1
redis==4.5.4
psycopg2-binary==2.9.5
qrcode==7.4.2
numpy==1.24.1
opencv-python==4.8.0.76
Pillow==9.3.0
